// src/pages/Home/queries.generated.ts
export type Maybe<T> = T | null;
// ...
export type BlogPostFragment = { title?: string | null, link?: string | null, categories: Array<string>, featuredImage?: { id: string, fileName: string, homeBlog?: string | null } | null };

export type LatestPostsQueryVariables = Types.Exact<{
  baseUrl: Types.Scalars['String'];
}>;


export type LatestPostsQuery = { blog: { url?: string | null, posts: Array<{ id: string, title?: string | null, link?: string | null, categories: Array<string>, featuredImage?: { id: string, fileName: string, homeBlog?: string | null } | null }> } };